Initial fishing block process

+++ Maple2.Server.Game/Util/RotationDirectionUtil.cs
@@ -7,14 +7,14 @@ using System.Xml;
 namespace Maple2.Server.Game.Util;
 
 public static class RotationDirectionUtil {
-    public const int SOUTH_EAST = 0;
-    public const int NORTH_EAST = 90;
-    public const int NORTH_WEST = 180;
-    public const int SOUTH_WEST = 270;
+    public const float SOUTH_EAST = 0;
+    public const float NORTH_EAST = 90;
+    public const float NORTH_WEST = 180;
+    public const float SOUTH_WEST = 270;
 
-    public static int GetClosestDirection(Vector3 rotation)
+    public static float GetClosestDirection(Vector3 rotation)
     {
-        int[] directions = new int[4]
+        float[] directions = new float[4]
         {
             SOUTH_EAST, NORTH_EAST, NORTH_WEST, SOUTH_WEST
         };
